refactor: 优化逆地理编码与版本号类型
PoiSearchHelper 优化：
- latLngToAddress 回调改为返回 (name, fullAddress) 双值
- 新增 extractShortName() 提取简洁地址名称
- 优先级：POI名称 > 街道门牌 > 区后部分 > 完整地址

MainFragment 适配：
- 使用新的双返回值接口，简化代码逻辑
- 移除原生 Geocoder，统一使用高德异步 API
- 提升代码可维护性和一致性

UpdateChecker & UpdateInfo 修复：
- versionCode 类型从 Int 改为 Long，避免溢出
- 使用 longVersionCode 替代 versionCode，兼容 Android 新规范
- 确保未来 versionCode 超过 21 亿时正常工作

+++ app/src/main/java/com/mockloc/util/UpdateInfo.kt
@@ -10,7 +10,7 @@ import java.io.Serializable
  */
 data class UpdateInfo(
     @SerializedName("versionCode")
-    val versionCode: Int,
+    val versionCode: Long,
     
     @SerializedName("versionName")
     val versionName: String,
@@ -25,7 +25,7 @@ data class UpdateInfo(
     val forceUpdate: Boolean = false,
     
     @SerializedName("minVersionCode")
-    val minVersionCode: Int = 1,
+    val minVersionCode: Long = 1L,
     
     @SerializedName("fileSize")
     val fileSize: Long = 0L,
@@ -40,14 +40,14 @@ data class UpdateInfo(
     /**
      * 判断是否有新版本
      */
-    fun hasUpdate(currentVersionCode: Int): Boolean {
+    fun hasUpdate(currentVersionCode: Long): Boolean {
         return versionCode > currentVersionCode
     }
     
     /**
      * 判断是否强制更新
      */
-    fun isForceUpdate(currentVersionCode: Int): Boolean {
+    fun isForceUpdate(currentVersionCode: Long): Boolean {
         return forceUpdate && versionCode > currentVersionCode
     }
